requirements.txt
python-socketio[client]
python-socketio
requests
eventlet